backend/requirements.txt
Flask
plaid-python
python-dotenv
flask-cors
flask_sqlalchemy
flask_login